+++ epyk/core/css/styles/GrpClsImage.py
@@ -1,5 +1,6 @@
 
 from epyk.core.css.styles import GrpCls
+from epyk.core.css import Classes
 from epyk.core.css.styles.attributes import AttrClsImage
 
 
@@ -21,3 +22,38 @@ class ClassIcon(GrpCls.ClassHtml):
     if self._css_struct is None:
       self._css_struct = AttrClsImage.AttrIcon(self.component)
     return self._css_struct
+
+
+class ClassTinySlider(GrpCls.ClassHtml):
+
+  def __init__(self, component):
+    super(ClassTinySlider, self).__init__(component)
+    self._css_tns_button, self._css_tns_active = None, None
+    self.classList['other'].add(self.cls_tns_button)
+    self.classList['other'].add(self.cls_tns_active)
+
+  @property
+  def cls_tns_button(self):
+    """
+    Description:
+    ------------
+
+    :rtype: Classes.CatalogImg.CatalogImg
+    """
+    if self._css_tns_button is None:
+      self._css_tns_button = Classes.CatalogImg.CatalogImg(
+        self.component.page, self.classList['other']).tns_button()
+    return self._css_tns_button
+
+  @property
+  def cls_tns_active(self):
+    """
+    Description:
+    ------------
+
+    :rtype: Classes.CatalogImg.CatalogImg
+    """
+    if self._css_tns_active is None:
+      self._css_tns_active = Classes.CatalogImg.CatalogImg(
+        self.component.page, self.classList['other']).tns_button_active()
+    return self._css_tns_active